refactor: wp.php into seperate classes
